Rearrange DAGS
