+++ dags/onet_extract.py
@@ -1,10 +1,9 @@
 """
-Workflow to create skills, job titles, and labeled corpora based on ONET data
-and common schema job listings
+Workflow to create skills, job titles, and importances based on ONET data
 
 """
 import csv
-import logging
+import io
 import os
 from datetime import datetime
 
@@ -13,16 +12,15 @@ from airflow.hooks import S3Hook
 from airflow.operators import BaseOperator
 
 from config import config
-from datasets import job_postings, OnetCache
-from utils.airflow import datetime_to_quarter
-from utils.s3 import upload
+from datasets import OnetCache
+from utils.es import basic_client
 from utils.hash import md5
+from utils.s3 import split_s3_path, upload
 
-from algorithms.corpus_creators.basic import SimpleCorpusCreator
-from algorithms.skill_taggers.simple import SimpleSkillTagger
 from algorithms.skill_extractors.onet_ksas import OnetSkillExtractor
 from algorithms.skill_importance_extractors.onet import OnetSkillImportanceExtractor
 from algorithms.title_extractors.onet import OnetTitleExtractor
+from algorithms.elasticsearch_indexers.job_titles_master import JobTitlesMasterIndexer
 
 # some DAG args, please tweak for sanity
 default_args = {
@@ -31,7 +29,7 @@ default_args = {
 }
 
 dag = DAG(
-    'simple_machine_learning',
+    'onet_extract',
     schedule_interval=None,
     default_args=default_args
 )
@@ -68,30 +66,6 @@ class TitleExtractOperator(BaseOperator):
         upload(conn, titles_filename, config['output_tables']['s3_path'])
 
 
-class SkillTagOperator(BaseOperator):
-    def execute(self, context):
-        conn = S3Hook().get_conn()
-        quarter = datetime_to_quarter(context['execution_date'])
-        labeled_filename = 'labeled_corpora_a'
-        with open(labeled_filename, 'w') as outfile:
-            writer = csv.writer(outfile, delimiter='\t')
-            job_postings_generator = job_postings(conn, quarter)
-            corpus_generator = SimpleCorpusCreator()\
-                .raw_corpora(job_postings_generator)
-            tagged_document_generator = \
-                SimpleSkillTagger(
-                    skills_filename=skills_filename,
-                    hash_function=md5
-                ).tagged_documents(corpus_generator)
-            for document in tagged_document_generator:
-                writer.writerow([document])
-        logging.info('Done tagging skills to %s', labeled_filename)
-        upload(
-            conn,
-            labeled_filename,
-            '{}/{}'.format(config['labeled_postings'], quarter)
-        )
-
 class SkillImportanceOperator(BaseOperator):
     def execute(self, context):
         conn = S3Hook().get_conn()
@@ -103,9 +77,26 @@ class SkillImportanceOperator(BaseOperator):
         skill_extractor.run()
         upload(conn, skill_importance_filename, config['output_tables']['s3_path'])
 
+
+class IndexJobTitlesMasterOperator(BaseOperator):
+    def execute(self, context):
+        conn = S3Hook()
+        input_bucket, input_prefix = split_s3_path(config['output_tables']['s3_path'])
+        key = conn.get_key(
+            '{}/{}'.format(input_prefix, titles_filename),
+            bucket_name=input_bucket
+        )
+        text = key.get_contents_as_string().decode('utf-8')
+        reader = csv.DictReader(io.StringIO(text), delimiter='\t')
+        JobTitlesMasterIndexer(
+            s3_conn=conn.get_conn(),
+            es_client=basic_client(),
+            job_title_generator=reader
+        ).replace()
+
 skills = SkillExtractOperator(task_id='skill_extract', dag=dag)
 titles = TitleExtractOperator(task_id='title_extract', dag=dag)
-skill_tagger = SkillTagOperator(task_id='skill_tag', dag=dag)
 skill_importance = SkillImportanceOperator(task_id='skill_importance', dag=dag)
+index_titles = IndexJobTitlesMasterOperator(task_id='job_titles_master', dag=dag)
 
-skill_tagger.set_upstream(skills)
+index_titles.set_upstream(titles)
